+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @openfn/cli
 
+## 1.13.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/project@0.2.0
+
 ## 1.13.4
 
 ### Patch Changes